* Remove defaults for bucket and prefix, they are now required parameters. * Allow uploading to be disabled with `-n`.

+++ scripts/repo-sync.sh
@@ -3,25 +3,32 @@ set -euo pipefail
 
 # Default variables.
 DEST=$(pwd)
-BUCKET=hoot-rpm
-KEEP=5
-PREFIX=develop
+BUCKET=""
+KEEP=10
+PREFIX=""
+UPLOAD=yes
 USAGE=no
 
 # These variables default to value of AWS CLI environment
 # variables (if defined).
 set +u
-PROFILE=${AWS_PROFILE:-hoot-rpm-develop}
+PROFILE=${AWS_PROFILE:-default}
 REGION=${AWS_DEFAULT_REGION:-us-east-1}
 set -u
 
-while getopts ":a:b:d:k:p:r:" opt; do
+# Getting parameters from the command line.
+while getopts ":a:b:p:d:k:nr:" opt; do
     case "${opt}" in
-        a)
-            PROFILE="${OPTARG}"
-            ;;
+        # Required parameters.
         b)
             BUCKET="${OPTARG}"
+            ;;
+        p)
+            PREFIX="${OPTARG}"
+            ;;
+        # Optional parameters.
+        a)
+            PROFILE="${OPTARG}"
             ;;
         d)
             DEST="${OPTARG}"
@@ -29,8 +36,8 @@ while getopts ":a:b:d:k:p:r:" opt; do
         k)
             KEEP="${OPTARG}"
             ;;
-        p)
-            PREFIX="${OPTARG}"
+        n)
+            UPLOAD="no"
             ;;
         r)
             REGION="${OPTARG}"
@@ -42,14 +49,17 @@ while getopts ":a:b:d:k:p:r:" opt; do
 done
 shift $((OPTIND-1))
 
-# Abort if invalid options.
-if [ "${USAGE}" == "yes" ]; then
-    echo "repo_sync.sh: [-a <AWS Profile>] [-b <S3 Bucket>] [-d <Local Destination>] [-p <S3 Prefix>] [-r <AWS Region>]"
+# Setting up.
+function usage() {
+    echo "repo_sync.sh: -b <S3 Bucket> -p <S3 Prefix> [-a <AWS CLI Profile>] [-d <Local Destination>] [-k <Keep>] [-n No Upload] [-r <AWS Region>]"
     exit 1
+}
+
+# Abort if invalid options.
+if [[ "${USAGE}" == "yes" || -z "${BUCKET}" || -z "${PREFIX}" ]]; then
+    usage
 fi
 
-
-## Setting up.
 REPO=$DEST/$PREFIX
 S3_URL="s3://${BUCKET}/${PREFIX}"
 if [ "${REGION}" == "us-east-1" ]; then
@@ -67,11 +77,13 @@ REPO_TS=$(aws s3api head-object --bucket $BUCKET --key $PREFIX/repodata/repomd.x
 
 if [ "${REPO_TS}" == "none" ]; then
     # Initialize the yum repository, as it doesn't exist in S3.
-    createrepo --database --unique-md-filenames --deltas $REPO
+    if [ ! -f $REPO/repodata/repomd.xml ] ; then
+        createrepo --database --unique-md-filenames --deltas $REPO
+    fi
 
     # Ensure a hoot.repo exists for use with yum-config-manager.
-    echo "cat > $REPO/hoot.repo <<EOF"
-    cat > $REPO/hoot.repo <<EOF
+    if [ ! -f $REPO/hoot.repo ]; then
+        cat > $REPO/hoot.repo <<EOF
 [hoot-develop]
 name = Hootenanny Development
 baseurl = https://${S3_HOST}/${BUCKET}/${PREFIX}
@@ -86,17 +98,20 @@ gpgcheck = 1
 repo_gpgcheck = 1
 gpgkey = https://s3.amazonaws.com/hoot-repo/el7/hoot.gpg
 EOF
+    fi
 else
-    # The repo already exists, sync the repository data, but keep
-    # any newer files/rpms.
+    # The repo already exists, sync the repository files down, but
+    # keep local files that are more recent than what's in S3.
     aws s3 sync $S3_URL $REPO --profile $PROFILE --region $REGION --keep-newer
 fi
 
-# Only keep as many packages as specified.
+# Only keep as many packages as specified, deleting older versions.
 repomanage --keep $KEEP --old $REPO | xargs rm -f -v
 
 # Update the repository metadata.
 createrepo --update $REPO
 
-# Sync back to the mothership, deleting
-aws s3 sync $REPO $S3_URL --delete --profile $PROFILE --region $REGION
+if [ "${UPLOAD}" == "yes" ]; then
+    # Upload back to the S3 bucket, deleting any files not present locally.
+    aws s3 sync $REPO $S3_URL --delete --profile $PROFILE --region $REGION
+fi
